chore: Move parseCsv into local module
