+++ modules/postman-templating/package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@types/lodash": "^4.14.155",
+    "csv-parse": "^4.10.1",
     "lodash": "^4.17.15",
     "squirrelly": "=8.0.0-beta.10",
     "xss": "^1.0.6"
