+++ modules/postman-templating/src/index.ts
@@ -1,2 +1,4 @@
 export * from './template-client'
 export * from './errors'
+export * from './parse-csv'
+export * from './types'
