+++ modules/postman-templating/src/types.ts
@@ -1,0 +1,1 @@
+export type CSVParams = { [key: string]: string }
